refactor: make HttpAuthenticationGateway implements SigninGateway and SignupGateway instead of CreateUserGateway
