+++ src/@core/frontend/domain/gateways/auth/signinGateway.ts
@@ -2,6 +2,6 @@ import { UserProps } from "@/@core/shared/entities/user/user"
 
 export type AuthData = Pick<UserProps, "email" | "password">
 
-export interface AuthenticationGateway {
-    authenticateUser(authData: AuthData): Promise<string | null>
+export interface SigninGateway {
+    signIn(data: AuthData): Promise<string | null>
 }
