+++ src/@core/frontend/main/gateways/authenticationGatewayFactory.ts
@@ -1,0 +1,7 @@
+import { HttpAuthenticationGateway } from "../../infra/gateways/auth/httpAuthenticationGateway"
+
+function createHttpAuthenticationGateway(apiUrl: string): HttpAuthenticationGateway {
+    return new HttpAuthenticationGateway(apiUrl)
+}
+
+export const httpAuthenticationGateway = createHttpAuthenticationGateway("/api")
