+++ src/@core/frontend/main/usecases/auth/signupFactory.ts
@@ -1,11 +1,8 @@
-import { HttpUserGateway } from "@/@core/frontend/infra/gateways/user/httpUserGateway"
+import { httpAuthenticationGateway } from "../../gateways/authenticationGatewayFactory"
 import { SignupUseCase } from "@/@core/frontend/usecases/auth/signupUseCase"
 
 function createSignupUseCase(): SignupUseCase {
-    const httpUserGateway = new HttpUserGateway()
-    const signupUseCase = new SignupUseCase(httpUserGateway)
-
-    return signupUseCase
+    return new SignupUseCase(httpAuthenticationGateway)
 }
 
 export const signupUseCase = createSignupUseCase()
