+++ src/@core/frontend/usecases/auth/signinUseCase.ts
@@ -1,13 +1,13 @@
 import { UserProps } from "@/@core/shared/entities/user/user"
-import { AuthenticationGateway } from "../../domain/gateways/auth/authenticationGateway"
+import { SigninGateway } from "../../domain/gateways/auth/signinGateway"
 
 type SigninInputDTO = Pick<UserProps, "email" | "password">
 
 export class SigninUseCase {
-    constructor(private readonly authenticationGateway: AuthenticationGateway) {}
+    constructor(private readonly signinGateway: SigninGateway) {}
 
     public async execute({ email, password }: SigninInputDTO): Promise<string | null> {
-        const accessToken = await this.authenticationGateway.authenticateUser({
+        const accessToken = await this.signinGateway.signIn({
             email,
             password
         })
