+++ src/@core/frontend/usecases/auth/signupUseCase.ts
@@ -1,15 +1,15 @@
 import { UserProps } from "@/@core/shared/entities/user/user"
-import { CreateUserGateway } from "../../domain/gateways/user/createUserGateway"
+import { SignupGateway } from "../../domain/gateways/auth/signupGateway"
 
 interface SignupInputDTO extends Omit<UserProps, "profileImg" | "phone"> {
     passwordConfirmation: string
 }
 
 export class SignupUseCase {
-    constructor(private readonly createUserGateway: CreateUserGateway) {}
+    constructor(private readonly signupGateway: SignupGateway) {}
 
     public async execute(signupData: SignupInputDTO): Promise<boolean> {
-        const isUserCreated = await this.createUserGateway.create(signupData)
+        const isUserCreated = await this.signupGateway.signUp(signupData)
 
         return isUserCreated
     }
